Wire Reliability network layer into hydra-node
